+++ curve25519/src/armv7/curve25519_armv7.c
@@ -1,0 +1,31 @@
+/*
+ * Copyright (C) 2023 - 2024, Stephan Mueller <smueller@chronox.de>
+ *
+ * License: see LICENSE file in root directory
+ *
+ * THIS SOFTWARE IS PROVIDED ``AS IS'' AND ANY EXPRESS OR IMPLIED
+ * WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES
+ * OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE, ALL OF
+ * WHICH ARE HEREBY DISCLAIMED.  IN NO EVENT SHALL THE AUTHOR BE
+ * LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
+ * CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT
+ * OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR
+ * BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF
+ * LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+ * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE
+ * USE OF THIS SOFTWARE, EVEN IF NOT ADVISED OF THE POSSIBILITY OF SUCH
+ * DAMAGE.
+ */
+
+#include "cpufeatures.h"
+#include "curve25519_armv7.h"
+#include "x25519_scalarmult.h"
+#include "x25519_scalarmult_c.h"
+
+int crypto_scalarmult_curve25519(unsigned char *q, const unsigned char *n,
+				 const unsigned char *p)
+{
+	if (lc_cpu_feature_available() & LC_CPU_FEATURE_ARM_NEON)
+		return crypto_scalarmult_curve25519_armv7(q, n, p);
+	return crypto_scalarmult_curve25519_c(q, n, p);
+}